+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - [TN] [ITN] [ta] Add Tamil text normalization and inverse text normalization grammars
 - [TN] [ITN] [te] Add Telugu text normalization and inverse text normalization grammars
 - [TN] [ITN] [ml] Add Malayalam text normalization and inverse text normalization grammars
+- [TN] [ITN] [kn] Add Kannada text normalization and inverse text normalization grammars